pyproject.toml
[tool.poetry]
name = "honeycomb"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djlint = "^1.35.2"
django-htmx = "^1.19.0"
pre-commit = "^3.8.0"
watchdog = "^5.0.2"
django-debug-toolbar = "^4.4.6"
django-environ = "^0.11.2"
anthropic = "^0.34.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
